virt-api: move functions near to their usage
These two functions are only used in the dialers file.
Move them there.
